move ApplicationModel from module rpc-api to module common
