+++ dubbo-common/src/main/java/org/apache/dubbo/common/model/ProviderModel.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.dubbo.rpc.model;
+package org.apache.dubbo.common.model;
 
 import java.util.Set;
